requirements.txt
fastapi
odbc
pandas
uvicorn
PyJWT